# FitnessArena-API/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "FitnessArena-API, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/programs/workouts/{id}": {
      "get": {
        "tags": [
          "programs"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/users/{id}": {
      "get": {
        "tags": [
          "users"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/user"
            }
          }
        }
      },
      "put": {
        "tags": [
          "users"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/user"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "users"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/users/signup": {
      "post": {
        "tags": [
          "users"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/user"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/user"
            }
          }
        }
      }
    },
    "/api/users/login": {
      "post": {
        "tags": [
          "users"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/login"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/users/image/{id}": {
      "post": {
        "tags": [
          "users"
        ],
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "formData",
            "name": "img",
            "type": "file"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/users/changepassword/{id}": {
      "post": {
        "tags": [
          "users"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/changePass"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/users/clients/{id}": {
      "get": {
        "tags": [
          "users"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/user"
              }
            }
          }
        }
      }
    },
    "/api/workouts": {
      "get": {
        "tags": [
          "workouts"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/workout"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "workouts"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/workout"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/workout"
            }
          }
        }
      }
    },
    "/api/workouts/{id}": {
      "get": {
        "tags": [
          "workouts"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/workout"
            }
          }
        }
      },
      "put": {
        "tags": [
          "workouts"
        ],
        "consumes": [
          "application/json-patch+json",
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/workout"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "workouts"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/workouts/traineeworkout/{id}": {
      "get": {
        "tags": [
          "workouts"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/programWorkout"
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
...
    "changePass": {
      "type": "object"
    },
...
    "login": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
...
    "programWorkout": {
      "type": "object",
      "properties": {
        "workoutId": {
          "format": "int32",
          "type": "integer"
        },
        "programId": {
          "format": "int32",
          "type": "integer"
        },
        "progWork": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
...
    "user": {
      "required": [
        "email",
        "fname",
        "gender",
        "lname",
        "password",
        "type"
      ],
      "type": "object",
      "properties": {
        "userId": {
          "format": "int32",
          "type": "integer"
        },
        "type": {
          "maxLength": 10,
          "minLength": 0,
          "type": "string"
        },
        "birthDate": {
          "format": "date-time",
          "type": "string"
        },
        "fname": {
          "maxLength": 20,
          "minLength": 0,
          "type": "string"
        },
        "lname": {
          "maxLength": 20,
          "minLength": 0,
          "type": "string"
        },
        "address": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "registerDate": {
          "format": "date-time",
          "type": "string"
        },
        "age": {
          "format": "int32",
          "type": "integer"
        },
        "gender": {
          "maxLength": 10,
          "minLength": 0,
          "type": "string"
        },
        "rating": {
          "format": "double",
          "type": "number"
        },
        "photoSrc": {
          "maxLength": 100,
          "minLength": 0,
          "type": "string"
        },
        "categoryName": {
          "maxLength": 20,
          "minLength": 0,
          "type": "string"
        },
        "target": {
          "maxLength": 20,
          "minLength": 0,
          "type": "string"
        },
        "cardNumber": {
          "format": "int32",
          "type": "integer"
        },
        "email": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        },
        "password": {
          "maxLength": 50,
          "minLength": 0,
          "type": "string"
        }
      }
    },
    "workout": {
      "required": [
        "targetMuscle"
      ],
      "type": "object",
      "properties": {
        "workoutId": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "maxLength": 10,
          "minLength": 0,
          "type": "string"
        },
        "targetMuscle": {
          "maxLength": 10,
          "minLength": 0,
          "type": "string"
        },
        "description": {
          "maxLength": 80,
          "minLength": 0,
          "type": "string"
        }
      }
    }
  }
}